% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transomics2cytoscape.R
\name{create3Dcyjs}
\alias{create3Dcyjs}
\title{Write cyjs file for transomics 3D visualization}
\usage{
create3Dcyjs(pathwayList, kinase2enzyme, outputcyjs)
}
\arguments{
\item{pathwayList}{a list with KEGG pathway ID and the Z height in 3D space}

\item{kinase2enzyme}{Tsv file path for transomic interaction edges}

\item{outputcyjs}{Output cyjs file path}
}
\value{
This function writes a cyjs file and returns no value.
}
\description{
Import multiple KEGG pathways and integrate the pathways
into a cyjs file for Cy3D renderer
}
\examples{
kinase2enzyme <- system.file('extdata', 'kinase_enzyme.txt',
                            package = 'transomics2cytoscape')
create3Dcyjs(c(rno00010=1, rno00010=200, rno04910=400),
             kinase2enzyme, 'transomics3D')
}
\author{
Kozo Nishida
}
