% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transomics2cytoscape.R
\name{createTransomicEdges}
\alias{createTransomicEdges}
\title{Create Trans-Omic edges between layers of the network.}
\usage{
createTransomicEdges(suid, transomicEdges)
}
\arguments{
\item{suid}{A SUID of Cytoscape network}

\item{transomicEdges}{Path of a TSV file with the 9 columns
(layer index of a source node,
name or KEGG object ID that the source node should have,
layer index of a target node,
name or KEGG object ID that the target node should have,
interaction type).}
}
\value{
A SUID of the 3D network.
}
\description{
Create Trans-Omic edges between layers of the network
}
\examples{
\dontrun{
transomicEdges <- system.file("extdata", "allosteric.tsv",
    package = "transomics2cytoscape")
createTransomicEdges(suid, transomicEdges)
}
}
\author{
Kozo Nishida
}
