% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transomics2cytoscape.R
\name{ec2reaction}
\alias{ec2reaction}
\title{Convert KEGG enzyme IDs to KEGG reaction IDs.}
\usage{
ec2reaction(tsvFilePath, columnIndex, outputFilename)
}
\arguments{
\item{tsvFilePath}{Path of a TSV file with the 9 columns
(layer index of a source node,
name or KEGG object ID that the source node should have,
layer index of a target node,
name or KEGG object ID that the target node should have,
interaction type).}

\item{columnIndex}{The column number}

\item{outputFilename}{The output filename}
}
\value{
None
}
\description{
Convert KEGG enzyme IDs to KEGG reaction IDs
}
\examples{
\dontrun{
ec = system.file("extdata", "allosteric_ecnumber.tsv",
    package = "transomics2cytoscape")
ec2reaction(ec, 8, "allosteric.tsv")
}
}
\author{
Kozo Nishida
}
