% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transomics2cytoscape.R
\name{installCyApps}
\alias{installCyApps}
\title{Install the Cytoscape Apps the transomics2cytoscape depends.}
\usage{
installCyApps()
}
\value{
None
}
\description{
Install the Cytoscape Apps the transomics2cytoscape depends
}
\examples{
\dontrun{
installCyApps()
}
}
\author{
Kozo Nishida
}
