% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transomics2cytoscape.R
\name{create3Dcyjs}
\alias{create3Dcyjs}
\title{Write cyjs file for transomics 3D visualization}
\usage{
create3Dcyjs(
  pathwayID1,
  pathwayID2,
  pathwayID3,
  zheight1,
  zheight2,
  zheight3,
  kinase2enzyme,
  outputcyjs
)
}
\arguments{
\item{pathwayID1}{a KEGG pathway ID}

\item{pathwayID2}{a KEGG pathway ID}

\item{pathwayID3}{a KEGG pathway ID}

\item{zheight1}{Z height for pathwayID1}

\item{zheight2}{Z height for pathwayID2}

\item{zheight3}{Z height for pathwayID3}

\item{kinase2enzyme}{Tsv file path for transomic interaction edges}

\item{outputcyjs}{Output cyjs file path}
}
\description{
Import multiple KEGG pathways and integrate the pathways
into a cyjs file for Cy3D renderer
}
\examples{
library(transomics2cytoscape)
kinase2enzyme <- system.file("extdata", "kinase_enzyme.txt",
                            package = "transomics2cytoscape")
create3Dcyjs("rno00010", "rno00010", "rno04910", 1, 200, 400,
             kinase2enzyme, "transomics3D.cyjs")
}
\author{
Kozo Nishida
}
