% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transomics2cytoscape.R
\name{create3Dnetwork}
\alias{create3Dnetwork}
\title{Create 3D network view for transomics visualization.}
\usage{
create3Dnetwork(networkLayers, transomicEdges, stylexml)
}
\arguments{
\item{networkLayers}{A TSV file path for the 3 column rows of 
(layer index, the network file path, Z height of the network layer)
in 3D space.}

\item{transomicEdges}{A TSV file path for the first 5 column rows of
(layer index of a source node, a name of source node,
layer index of a target node, a name of target node,
interaction type)
representing transomic interaction edges between the network layers.}

\item{stylexml}{A XML file path for Cytoscape style file to be applied to
the 3D network.}
}
\value{
A SUID of the 3D network.
}
\description{
Import multiple KEGG pathways and integrate the pathways
into Cy3D renderer
}
\examples{
\donttest{
sif <- system.file("extdata","galFiltered.sif",package="RCy3")
file.copy(sif, ".")
networkLayers <- system.file("extdata", "networkLayers.tsv",
    package = "transomics2cytoscape")
transomicEdges <- system.file("extdata", "transomicEdges.tsv",
    package = "transomics2cytoscape")
stylexml <- system.file("extdata", "transomics.xml",
    package = "transomics2cytoscape")
create3Dnetwork(networkLayers, transomicEdges, stylexml)
}
}
\author{
Kozo Nishida
}
