% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transomics2cytoscape.R
\name{create3Dnetwork}
\alias{create3Dnetwork}
\title{Create 3D network view for transomics visualization.}
\usage{
create3Dnetwork(pathwayZheights, kinase2enzyme, stylexml)
}
\arguments{
\item{pathwayZheights}{A list with KEGG pathway ID and Z height in 3D
space.}

\item{kinase2enzyme}{A TSV file path for the edges between pathway network
layers.}

\item{stylexml}{A XML file path for Cytoscape style file to be applied to
the 3D network.}
}
\value{
A SUID of the 3D network.
}
\description{
Import multiple KEGG pathways and integrate the pathways
into Cy3D renderer
}
\examples{
\donttest{
kinase2enzyme <- system.file("extdata", "kinase_enzyme.txt",
    package = "transomics2cytoscape")
stylexml <- system.file("extdata", "transomics.xml",
    package = "transomics2cytoscape")
create3Dnetwork(c(rno00010=1, rno00010=200, rno04910=400, rno04910=600),
    kinase2enzyme, stylexml)
}
}
\author{
Kozo Nishida
}
