% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transomics2cytoscape.R
\name{create3Dnetwork}
\alias{create3Dnetwork}
\title{Create 3D network view for transomics visualization.}
\usage{
create3Dnetwork(networkDataDir, networkLayers, transomicEdges, stylexml)
}
\arguments{
\item{networkDataDir}{Path of a directory to put the network files
of the second column of networkLayers TSV.}

\item{networkLayers}{Path of a TSV file with the 3 columns (layer index,
the network file name in networkDataDir, Z-height of the network).}

\item{transomicEdges}{Path of a TSV file with the 5 columns
(layer index of a source node,
name or KEGG object ID that the source node should have,
layer index of a target node,
name or KEGG object ID that the target node should have,
interaction type).}

\item{stylexml}{Path of a XML file for Cytoscape style}
}
\value{
A SUID of the 3D network.
}
\description{
Import multiple KEGG pathways and integrate the pathways
into Cy3D renderer
}
\examples{
\donttest{
networkDataDir <- tempfile(); dir.create(networkDataDir)
sif <- system.file("extdata","galFiltered.sif",package="RCy3")
file.copy(sif, networkDataDir)
networkLayers <- system.file("extdata", "networkLayers.tsv",
    package = "transomics2cytoscape")
transomicEdges <- system.file("extdata", "transomicEdges.tsv",
    package = "transomics2cytoscape")
stylexml <- system.file("extdata", "transomics.xml",
    package = "transomics2cytoscape")
create3Dnetwork(networkDataDir, networkLayers, transomicEdges, stylexml)
}
}
\author{
Kozo Nishida
}
