% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transomics2cytoscape.R
\name{ec2reaction}
\alias{ec2reaction}
\title{Convert KEGG enzyme IDs to KEGG reaction IDs.}
\usage{
ec2reaction(tsvFilePath, columnIndex, outputFilename)
}
\arguments{
\item{tsvFilePath}{Path of a TSV file with column of EC number}

\item{columnIndex}{Index number of the column with the EC number you want
to convert}

\item{outputFilename}{The output file name}
}
\value{
None
}
\description{
Convert a EC number column to a KEGG reaction ID column
}
\examples{
\dontrun{
layer3to2 <- system.file("extdata/usecase1", "allosteric_ecnumber.tsv",
    package = "transomics2cytoscape")
ec2reaction(layer3to2, 6, "allosteric_ec2rea.tsv")
}
}
\author{
Kozo Nishida
}
